requirements/base.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --output-file=requirements/base.txt pyproject.toml
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.6
    # via
    #   discord-py
    #   realtime
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via
    #   httpx
    #   openai
    #   starlette
async-lru==2.0.4
    # via redstone-squid (pyproject.toml)
attrs==24.3.0
    # via aiohttp
audioop-lts==0.2.1
    # via discord-py
beartype==0.19.0
    # via redstone-squid (pyproject.toml)
beautifulsoup4==4.12.3
    # via redstone-squid (pyproject.toml)
braceexpand==0.1.7
    # via jishaku
brotli==1.1.0
    # via discord-py
cachetools==5.5.0
    # via google-auth
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   jishaku
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
deprecated==1.2.15
    # via vecs
deprecation==2.1.0
    # via postgrest
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@8686b5d1bbc1d3c862292eb436ab630d6e9c9b53
    # via redstone-squid (pyproject.toml)
discord-py==2.5.2
    # via
    #   redstone-squid (pyproject.toml)
    #   discord-ext-menus
    #   jishaku
distro==1.9.0
    # via openai
fastapi==0.115.6
    # via redstone-squid (pyproject.toml)
flupy==1.2.1
    # via vecs
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via redstone-squid (pyproject.toml)
google-auth==2.37.0
    # via
    #   google-auth-oauthlib
    #   gspread
google-auth-oauthlib==1.2.1
    # via gspread
gotrue==2.11.1
    # via supabase
gspread==6.1.4
    # via redstone-squid (pyproject.toml)
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
h2==4.1.0
    # via httpx
hpack==4.0.0
    # via h2
httpcore==1.0.7
    # via httpx
httplib2==0.22.0
    # via oauth2client
httpx==0.27.2
    # via
    #   gotrue
    #   openai
    #   postgrest
    #   storage3
    #   supabase
    #   supafunc
hyperframe==6.0.1
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
import-expression==2.2.1.post1
    # via jishaku
jishaku==2.6.0
    # via redstone-squid (pyproject.toml)
jiter==0.8.2
    # via openai
markdown==3.7
    # via redstone-squid (pyproject.toml)
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
numpy==2.2.1
    # via pgvector
oauth2client==4.1.3
    # via redstone-squid (pyproject.toml)
oauthlib==3.2.2
    # via requests-oauthlib
openai==1.58.1
    # via redstone-squid (pyproject.toml)
orjson==3.10.14
    # via discord-py
packaging==24.2
    # via deprecation
pgvector==0.3.6
    # via vecs
postgrest @ git+https://github.com/redstone-squid/postgrest-py@acaa901cbfc294912a76179f0cc81c09eccd4c22
    # via
    #   redstone-squid (pyproject.toml)
    #   supabase
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
psycopg2-binary==2.9.10
    # via vecs
pyasn1==0.6.1
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via
    #   google-auth
    #   oauth2client
pydantic==2.10.4
    # via
    #   fastapi
    #   gotrue
    #   openai
    #   postgrest
pydantic-core==2.27.2
    # via pydantic
pyparsing==3.2.0
    # via httplib2
python-dateutil==2.9.0.post0
    # via
    #   realtime
    #   storage3
python-dotenv==1.0.1
    # via redstone-squid (pyproject.toml)
realtime==2.4.3
    # via supabase
requests==2.32.3
    # via requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9
    # via
    #   google-auth
    #   oauth2client
six==1.17.0
    # via
    #   oauth2client
    #   python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.6
    # via beautifulsoup4
sqlalchemy==2.0.36
    # via vecs
starlette==0.41.3
    # via fastapi
storage3==0.11.0
    # via supabase
strenum==0.4.15
    # via supafunc
supabase==2.15.2
    # via redstone-squid (pyproject.toml)
supafunc==0.9.0
    # via supabase
tabulate==0.9.0
    # via jishaku
tqdm==4.67.1
    # via openai
typing-extensions==4.13.2
    # via
    #   fastapi
    #   flupy
    #   jishaku
    #   openai
    #   pydantic
    #   pydantic-core
    #   realtime
    #   sqlalchemy
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via redstone-squid (pyproject.toml)
vecs==0.4.5
    # via redstone-squid (pyproject.toml)
websockets==13.1
    # via realtime
wrapt==1.17.0
    # via deprecated
yarl==1.18.3
    # via aiohttp
zstandard==0.23.0
    # via discord-py
